requirements.txt
pyshark==0.6.0
Pillow>=9.0.0
protobuf==5.27.2
psutil==5.9.5
typing-extensions>=4.0.0
python-dotenv>=0.19.0
pathlib>=1.0.1
